pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-vm-mcp"
dynamic = ["version"]
description = "MCP server for llm-nano-vm — run deterministic LLM programs via Model Context Protocol"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "llm-nano-vm>=0.5.0",
    "mcp>=1.0.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
nano-vm-mcp = "nano_vm_mcp.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nano_vm_mcp*"]

[tool.setuptools_scm]
write_to = "nano_vm_mcp/_version.py"
fallback_version = "0.1.0.dev0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E501", "E402", "E741", "F401", "F841"]
